migration_holder/pyproject.toml
[tool.poetry]
name = "drone-mob"
version = "0.2.0"
description = "Autonomous Man Overboard (MOB) detection and rescue drone system."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/theogodfrey/drone-mob"

[tool.poetry.dependencies]
python = "^3.9"

# Original dependencies
pyyaml = ">=6.0"
numpy = ">=1.21.0" # <-- Already here, but now critical for Probabilistic AI
scipy = ">=1.7.0"
opencv-python = ">=4.5.0" # <-- Already here, but now critical for Media Server

# Refactoring dependencies
pydantic = ">=2.0"
transitions = {version = ">=0.9.0", extras = ["async"]}
websockets = ">=10.0"
paho-mqtt = ">=1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

